Move miscellaneous case to more appropriate test file
