+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/keyboard/CutCopyPasteTests.java
@@ -1,14 +1,18 @@
 package org.fxmisc.richtext.keyboard;
 
 import com.nitorcreations.junit.runners.NestedRunner;
+import javafx.scene.Scene;
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
 import javafx.stage.Stage;
+import org.fxmisc.flowless.VirtualizedScrollPane;
+import org.fxmisc.richtext.CodeArea;
 import org.fxmisc.richtext.InlineCssTextAreaAppTest;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.testfx.framework.junit.ApplicationTest;
 
 import static javafx.scene.input.KeyCode.*;
 import static junit.framework.TestCase.assertEquals;
@@ -271,4 +275,35 @@ public class CutCopyPasteTests extends InlineCssTextAreaAppTest {
 
         }
     }
+
+    public class MiscellaneousCases extends ApplicationTest {
+
+        CodeArea area;
+
+        @Override
+        public void start(Stage primaryStage) throws Exception {
+            area = new CodeArea("abc\ndef\nghi");
+            VirtualizedScrollPane<CodeArea> vsPane = new VirtualizedScrollPane<>(area);
+
+            Scene scene = new Scene(vsPane, 400, 400);
+            primaryStage.setScene(scene);
+            primaryStage.show();
+        }
+
+        public class When_User_Makes_Selection_Ending_In_Newline_Character {
+
+            @Before
+            public void setup() {
+                area.selectRange(2, 4);
+            }
+
+            @Test
+            public void copying_and_pasting_should_not_throw_exception() {
+                push(CONTROL, C);
+
+                push(CONTROL, V);
+            }
+        }
+    }
+
 }
